macOS: set com.apple.quarantine xattr for TLFs whith other writers
Note that this breaks running *.app directly from such KBFS TLFs. I've
tried several things including modifying the
com.apple.LaunchServices.QuarantineEventsV2 sqlite DB, but I just can't
get macOS to pass the integrity check when the com.apple.quarantine
xattr is set. I think it might require setting xattr for stuff inside
the .app individually. Not sure. But this seems to work for scripts
pretty well at least. It triggers a prompt when user tries to run a
script in a TLF where we are not the only writer. If user chooses to
open it anyway, next time user tries to open it the prompt won't show.

This PR doesn't persist it yet though. In the next PR I'll have the
xattr stored in a local DB so it persists through restarts.